+++ docs/source/conf.py
@@ -20,9 +20,6 @@ import warnings
 import sphinx_rtd_theme
 from sqlalchemy.exc import SAWarning
 
-# set sample config when building docs
-os.environ['INDICO_CONFIG'] = os.path.join(os.path.dirname(__file__), '..', '..', 'indico', 'indico.conf.sample')
-
 # silence sqlalchemy SAWarning
 warnings.simplefilter('ignore', SAWarning)
 
@@ -46,6 +43,7 @@ extensions = ['sphinx.ext.autodoc',
               'sphinx.ext.todo',
               'sphinx.ext.coverage',
               'sphinx.ext.imgmath',
+              'sphinx.ext.intersphinx',
               'sphinx_issues',
               'exec_directive']
 
@@ -115,6 +113,12 @@ pygments_style = 'sphinx'
 # include __init__() as well
 autoclass_content = "both"
 autodoc_default_flags = ['show-inheritance']
+
+# Intersphinx urls
+intersphinx_mapping = {
+    'python': ('https://docs.python.org/2.7/', None),
+    'celery': ('http://docs.celeryproject.org/en/stable/', None),
+}
 
 # -- Options for HTML output ---------------------------------------------------