+++ indico/legacy/services/implementation/search.py
@@ -14,6 +14,8 @@
 # You should have received a copy of the GNU General Public License
 # along with Indico; if not, see <http://www.gnu.org/licenses/>.
 
+from __future__ import unicode_literals
+
 from itertools import chain
 
 from indico.core.db.sqlalchemy.custom.unaccent import unaccent_match
@@ -30,6 +32,7 @@ from indico.util.string import sanitize_email, to_unicode
 
 class SearchBase(LoggedOnlyService):
     CHECK_HTML = False
+    UNICODE_PARAMS = True
 
     def _process_args(self):
         self._searchExt = self._params.get('search-ext', False)
@@ -69,7 +72,6 @@ class SearchUsers(SearchBase):
 
 
 class SearchGroups(SearchBase):
-
     def _process_args(self):
         SearchBase._process_args(self)
         self._group = self._params.get("group", "").strip()