+++ CHANGES.rst
@@ -23,6 +23,8 @@ Bugfixes
   a number field
 - Fix "N/A" not being displayed in a survey result if nothing is
   selected in a multi-choice select field
+- Fix error when using ``target_*`` placeholders in abstract
+  notification emails for actions other than "Merged" (:issue:`3171`)
 
 
 Version 2.0rc1